Add Normal Equation Method to solve the theta
Normal Equation method does not need the parameter alpha and iteration process.

When the number of feature is very large, this method will be very slow.

When calculating the inverse matrix of X.T*X, we should use the "pinv" instead of "inv" in case that X.T*X is singular.

the reasons that X.T*X might be singular are mainly from two aspects. The first is feature redundant. The second is too many features but a little instances